promart/services/products_services/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-appconf==1.1.0
django-celery-email==3.0.0
django-elasticsearch-dsl==8.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
